requirements.txt
instapy
tweepy
slackclient
apache-airflow
celery
redis
coveralls
hive
codecov
pytest
#apache-airflow[postgres,mysql]